+++ utils/templates/schema/Elk_Aerial_SRB_Recruit_Comp_Survey_2.0_Validation_Config_Builder.ts
