+++ utils/templates/schema/Goat_2.0_Validation_Config_Builder.ts
