+++ utils/templates/schema/Moose_Aerial_Transect_Distance_Sampling_Survey_2.0_Validation_Config_Builder.ts
