+++ utils/templates/schema/validation-config-helpers.ts
